#60 value 타입 패키지 변경
