+++ src/main/java/toy/blog/be/controller/post/api/PostApiController.java
@@ -11,7 +11,7 @@ import toy.blog.be.controller.post.dto.request.PostCreateRequest;
 import toy.blog.be.controller.post.dto.request.PostUpdateRequest;
 import toy.blog.be.controller.post.dto.response.PostResponse;
 import toy.blog.be.domain.entity.Post;
-import toy.blog.be.domain.entity.PostId;
+import toy.blog.be.domain.value.PostId;
 import toy.blog.be.service.PostService;
 
 import javax.validation.Valid;
