+++ src/main/java/toy/blog/be/domain/entity/Keywords.java
@@ -4,11 +4,10 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import toy.blog.be.infra.IdGenerator;
+import toy.blog.be.domain.value.KeywordId;
 
 import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
-import javax.persistence.Id;
 import javax.persistence.Table;
 import java.time.LocalDateTime;
