+++ src/main/java/toy/blog/be/repository/KeywordRepository.java
@@ -1,7 +1,7 @@
 package toy.blog.be.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import toy.blog.be.domain.entity.KeywordId;
+import toy.blog.be.domain.value.KeywordId;
 import toy.blog.be.domain.entity.Keywords;
 
 import java.util.Optional;
