+++ src/main/java/toy/blog/be/repository/PostRepository.java
@@ -3,7 +3,7 @@ package toy.blog.be.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import toy.blog.be.domain.entity.Post;
-import toy.blog.be.domain.entity.PostId;
+import toy.blog.be.domain.value.PostId;
 
 @Repository
 public interface PostRepository extends JpaRepository<Post, PostId> {
