+++ src/main/java/toy/blog/be/service/PostService.java
@@ -8,10 +8,10 @@ import toy.blog.be.controller.PagedRequest;
 import toy.blog.be.controller.PagedResponse;
 import toy.blog.be.controller.post.dto.response.KeywordDTO;
 import toy.blog.be.controller.post.dto.response.PostResponse;
-import toy.blog.be.domain.entity.KeywordId;
+import toy.blog.be.domain.value.KeywordId;
 import toy.blog.be.domain.entity.Keywords;
 import toy.blog.be.domain.entity.Post;
-import toy.blog.be.domain.entity.PostId;
+import toy.blog.be.domain.value.PostId;
 import toy.blog.be.infra.IdGenerator;
 import toy.blog.be.repository.KeywordRepository;
 import toy.blog.be.repository.PostRepository;
